FirstCommit

+++ src/main/java/com/StudentManagement/service/impl/StudentServiceimpl.java
@@ -2,6 +2,7 @@ package com.StudentManagement.service.impl;
 
 import org.springframework.stereotype.Service;
 
+import com.StudentManagement.exception.IDNotFoundException;
 import com.StudentManagement.model.Student;
 import com.StudentManagement.repository.StudentRepository;
 import com.StudentManagement.service.StudentService;
@@ -27,29 +28,29 @@ public class StudentServiceimpl implements StudentService {
 
 	@Override
 	public Student displayById(Long id) {
-		/*
-		 * if(existsById(id)) { return studentRepository.findById(id).get(); }
-		 */
-		return studentRepository.findById(id).get();
+		if (studentRepository.findById(id).isPresent()) {
+			return studentRepository.findById(id).get();
+		} else {
+			throw new IDNotFoundException("Student", "Id", id);
+		}
 	}
 
 	@Override
 	public Student updateStudent(Long id) {
-		Student existingStudent = studentRepository.findById(id).get();
-		/*
-		 * existingStudent.setFName(updatedStudent.getFName());
-		 * existingStudent.setLName(updatedStudent.getLName());
-		 * existingStudent.setGender(updatedStudent.getGender());
-		 * existingStudent.setAddress(updatedStudent.getAddress());
-		 * existingStudent.setEmail(updatedStudent.getEmail());
-		 * existingStudent.setMobile(updatedStudent.getMobile());
-		 */
-		return existingStudent;
+		if (studentRepository.findById(id).isPresent()) {
+			return studentRepository.findById(id).get();
+		} else {
+			throw new IDNotFoundException("Student", "Id", id);
+		}
 	}
 
 	@Override
 	public void deleteStudentbyId(Long id) {
-		studentRepository.deleteById(id);
+		if (studentRepository.findById(id).isPresent()) {
+			studentRepository.deleteById(id);
+		} else {
+			throw new IDNotFoundException("Student", "Id", id);
+		}
 	}
 
 	@Override
